+++ public/admin/_app/immutable/nodes/14.CLnbJIWA.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/Ch6stX1y2.js";export{e as component};