+++ public/admin/_app/immutable/nodes/17.4EhtC5-b.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/BZ7qcN9k2.js";export{e as component};