+++ public/admin/_app/immutable/nodes/18.BHNyCmKw.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/BuWBY3Rr2.js";export{e as component};