+++ public/admin/_app/immutable/nodes/30.BE0ihj4P.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/C1PyGc6R.js";export{e as component};