+++ public/admin/_app/immutable/nodes/32.Cdj5ha-o.js
@@ -1,0 +1,1 @@
+import"../chunks/CpTjFNj2.js";import"../chunks/xihTtKlq.js";import"../chunks/CO3Brb12.js";import{t as e}from"../chunks/C1PyGc6R.js";function t(t){e(t,{})}export{t as component};